Remove unused configuration and component files
The unused configuration files have been deleted which include ".eslintrc.json" and ".yarnrc.yml". Likewise, the components which are not in use have also been deleted from the source code which includes "3d-card.tsx", "3d-pin.tsx", and "AboutSection.tsx" among others. This is done to declutter the source code and to maintain focus on the necessary components and configurations.