Update layout and add scrolling feature to homepage
The layout of the application has been updated with changes in typography and color scheme. In addition, a new scrolling feature has been added to the homepage which adjusts the text position according to the scroll position. These changes provide a better visual experience and improve user interaction with the application.